Support remote option

+++ crawl.py
@@ -7,13 +7,17 @@ from Classes import *
 parser = argparse.ArgumentParser(description='Crawler')
 parser.add_argument("--debug", action='store_true',  help="Dont use path deconstruction and recon scan. Good for testing single URL")
 parser.add_argument("--url", help="Custom URL to crawl")
+parser.add_argument("--remote", help="Remote WebDriver URL")
 args = parser.parse_args()
 
 # Clean form_files/dynamic
 root_dirname = os.path.dirname(__file__)
 dynamic_path = os.path.join(root_dirname, 'form_files', 'dynamic')
-for f in os.listdir(dynamic_path):
-    os.remove(os.path.join(dynamic_path, f))
+if os.path.exists(dynamic_path):
+    for f in os.listdir(dynamic_path):
+        os.remove(os.path.join(dynamic_path, f))
+else:
+    os.makedirs(dynamic_path)
 
 WebDriver.add_script = add_script
 
@@ -23,7 +27,10 @@ chrome_options.add_argument("--disable-web-security")
 chrome_options.add_argument("--disable-xss-auditor")
 
 # launch Chrome
-driver = webdriver.Chrome(options = chrome_options)
+if args.remote:
+    driver = webdriver.Remote(command_executor=args.remote, options=chrome_options)
+else:
+    driver = webdriver.Chrome(options = chrome_options)
